vaelsys_fashion/main.py
import torch
from torch.utils.data import Dataset, DataLoader
from torchvision import transforms
from PIL import Image
import deepface
from efficientnet_pytorch import EfficientNet

from torchvision import transforms, datasets
from sklearn.model_selection import train_test_split


print("Importing is done ...")