vaelsys_fashion/requirements.txt
uWSGI==2.0.23
torch==1.11.0
torchsummary==1.5.1
torchvision==0.12.0
distlib==0.3.4
dj-database-url==0.5.0
psycopg2>=2.8.6,<2.9
django-widget-tweaks==1.4.12
filelock==3.4.2
platformdirs==2.4.1
tzdata==2021.5
virtualenv==20.12.0
whitenoise==6.0.0
asgiref==3.6.0
Django==3.2.16
numpy
Pillow==9.4.0
opencv-python-headless
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
python-dotenv==1.0.0
tqdm==4.66
efficientnet-pytorch==0.7.1
deepface==0.0.90
